# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Blog Educacional API
  description: API REST para plataforma de blog educacional da FIAP
  version: 1.0.0
  contact:
    name: Vinicius Franco Silva
    email: aluno@fiap.com.br
servers:
  - url: http://localhost:3001
    description: Servidor de desenvolvimento local
  - url: http://localhost:3000
    description: Servidor de produção (Docker)
...
paths:
...
  /usuarios:
    get:
      tags:
        - Usuários
      summary: Listar usuários (somente admin)
      description: Retorna usuários paginados, com filtro opcional por papel. Exige papel admin.
      operationId: listarUsuarios
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: papel
          in: query
          schema:
            type: string
            enum: [docente, admin, aluno]
      responses:
        '200':
          description: Lista paginada de usuários
        '401':
          description: Não autenticado
        '403':
          description: Permissão insuficiente (requer admin)
    post:
      tags:
        - Usuários
      summary: Criar usuário (somente admin)
      description: Permite que um admin crie um usuário com papel arbitrário.
      operationId: criarUsuario
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, senha, nome]
              properties:
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  format: password
                  minLength: 8
                nome:
                  type: string
                  minLength: 2
                papel:
                  type: string
                  enum: [docente, admin, aluno]
                  default: docente
      responses:
        '201':
          description: Usuário criado
        '400':
          description: "Dados inválidos (ex.: senha fraca)"
        '401':
          description: Não autenticado
        '403':
          description: Permissão insuficiente (requer admin)
        '409':
          description: Email já está em uso
...
  /posts:
    get:
      tags:
        - Posts
      summary: Listar todos os posts
      description: Retorna lista de todos os posts ordenados por data de criação (mais recentes primeiro)
      operationId: listarPosts
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    post:
      tags:
        - Posts
      summary: Criar novo post
      description: Cria uma nova postagem (requer autenticação)
      operationId: criarPost
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarPostInput'
      responses:
        '201':
          description: Post criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroValidacao'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '403':
          description: Permissão insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido no endpoint /auth/login
  schemas:
...
    Post:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Introdução ao React
        content:
          type: string
          example: React é uma biblioteca JavaScript para construir interfaces de usuário...
        author:
          type: string
          example: João Silva
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z
    CriarPostInput:
      type: object
      required:
        - title
        - content
        - author
      properties:
        title:
          type: string
          minLength: 3
          example: Introdução ao React
        content:
          type: string
          minLength: 10
          example: React é uma biblioteca JavaScript para construir interfaces de usuário...
        author:
          type: string
          minLength: 2
          example: João Silva
...
    Erro:
      type: object
      properties:
        message:
          type: string
          example: Erro interno do servidor
    ErroValidacao:
      type: object
      properties:
        message:
          type: string
          example: Falha na validação dos campos
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email inválido